# backend/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: App Backend API
  version: 0.1.0
  description: |
    The template's API contract. MAINTAINED PAIRWISE with the Postman
    collection at docs/postman/ — any route/param/contract change updates both
    in the same commit. Served as Swagger UI at `<base>/docs/` when
    API_DOCS_ENABLED (on in the dev stack, off everywhere else).

    Authentication: the browser signs in through POST /auth/login and carries
    HttpOnly cookies; direct callers (this UI, Postman) send a Keycloak Bearer
    token instead. Every error is RFC-7807 problem+json.
servers:
  - url: /app/v1
    description: Frontend proxy path
  - url: /api/v1
    description: Direct path (Postman, smoke tests)
...
paths:
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes (search + paging)
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: A page of notes
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Note' }
                  total: { type: integer }
        '401': { $ref: '#/components/responses/Problem' }
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note (MANAGER/ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                body: { type: string }
                tags: { type: array, items: { type: string } }
                actionUuid:
                  type: string
                  description: Optional replay protection — the same uuid twice answers 409.
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /notes/{noteId}:
    parameters:
      - { name: noteId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: getNote
      tags: [notes]
      summary: One note
      responses:
        '200':
          description: The note
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Change title/body/tags (MANAGER/ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                body: { type: string }
                tags: { type: array, items: { type: string } }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note (MANAGER/ADMIN)
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
...
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: app_access
  responses:
    Problem:
      description: RFC-7807 problem+json
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Note:
      type: object
      properties:
        noteId: { type: string, format: uuid }
        title: { type: string }
        body: { type: string }
        tags: { type: array, items: { type: string } }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedBy: { type: [string, 'null'] }
        updatedAt: { type: [string, 'null'], format: date-time }
...
    Problem:
      type: object
      description: RFC-7807 problem+json — the error shape every endpoint answers with.
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        violations:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }